审核表添加单据动作

+++ app/src/main/java/cn/atsoft/dasheng/action/FormActionInterface.java
@@ -38,5 +38,5 @@ public interface FormActionInterface {
         documentsStatusService.save(documentsStatus);
 
     }
-    Long getStatus();
+    Integer getStatus();
 }
